Make Builder::message take self by mut reference (#9)
This changes the API to take Self by &mut reference. Practice has shown that the by-value Self is very unwieldy in this particular case, because any error that arises from the message method will consume both the builder and the messages built this far. While that could be rectified by returning Self in Err case, taking a &mut is a significantly more elegant solution.

* Fix documentation without features