Replace Travis with Github Actions
Github Actions is the preferred CI in the rust ecosystem nowadays,
especially after Travis has been acquired by Idera, a company known to
neglect their portfolio.

There are some benefits to using GHA – with it we can test the three
major platforms, the parallelism is significantly greater and the
integration is also better.