Implement a Sink interface for hedwig publishing
This introduces [`Sink`][1] implementations which validate and publish
messages to hedwig. See the docs for details.

The underlying implementation supports retries, but does not actually
perform them. I plan to add a retrying Sink wrapper at some point in the
future.

I also took the liberty to add a type alias for topics, because `&'static
str` is not particularly descriptive for a type.

[1]: https://docs.rs/futures/0.3.13/futures/sink/trait.Sink.html