Add traits for Consumers
This adds support for hedwig consumers -- or at least support for
implementations to be made. This includes traits for consumer streams,
message decoding, message acknowledgement, etc.

The concrete PubSub implementation currently depends on some private
crates, so it won't be included in this public repo until those
dependencies are open-sourced themselves, hopefully in the near future.