Fix appending the API_BODY_SUFFIX twice
When the message batch approaches the limit we will hit the branch
within which we accidentally appended the trailing `]}` twice to the
request body.

This branch was previously covered by our tests, in particular the
`segmenter_preserves_order_when_splitting`, but due to the lack of
assertions we never actually verified that the JSON we produce is valid.

Here we remove the invalid append, and some debug assertions as well as
a regression test for this issue.