Refactor Publishing (#34)
* Refactor Publishing

This changeset completely refactors publishing to be based on async
Sinks, and to port the googlepubsub publisher to use gRPC instead of the
JSON REST api. Together, these should make the library easier to use,
compose better with the broader rust async ecosystem, and provide better
performance for publishing.

This additionally simplifies the crate hierarchy and some of the crate
features, as the divisions only made much sense when publishing and
consuming were separate implementations. Now backends are expected to
provide publishing and consuming together at the same time.

Most of this code is adapted from the private crate standard-hedwig,
which has been a proving ground for the publishing API before it could
be incorporated into this open-source crate, and has seen effective
operation in production.

* Bump major version to 5